Update find first part algorithm
Previously, we were only looking at instances in the root assembly,
and an assembly could be assigned a body. Now, we look at all instances
that are not assemblies to assign bodies, so we need to recurse into
assemblies in the instances list and do a depth-first search to find
the first part that would appear in the instances tree in the Onshape
web app. We select this first part as the root body of the assembly.